*: Cleanup table codec
Prepare for row-based storage model
